cache: fix bugs, add logs

+++ src/cache.h
@@ -30,12 +30,12 @@ struct cache {
 extern int ppcemu_cache_init(struct _ppcemu_state *state, uint icache_size, uint dcache_size);
 extern void ppcemu_cache_destroy(struct _ppcemu_state *state);
 extern void ppcemu_icache_fetch(struct cache *icache, u32 addr, u32 *out);
-extern void ppcemu_icache_invalidate_line(struct cache *icache, u64 addr);
+extern void ppcemu_icache_invalidate_line(struct cache *icache, u32 addr);
 extern void ppcemu_dcache_load(struct cache *dcache, u32 addr, uint size, void *out);
 extern void ppcemu_dcache_store(struct cache *dcache, u32 addr, unsigned size, void *in);
 extern void ppcemu_dcache_writeback_line(struct cache *dcache, u32 addr);
 extern void ppcemu_dcache_invalidate_line(struct cache *dcache, u32 addr);
-extern void ppcemu_dcache_writeback_invalidate_line(struct cache *dcache, u64 addr);
+extern void ppcemu_dcache_writeback_invalidate_line(struct cache *dcache, u32 addr);
 extern void ppcemu_cache_invalidate_all(struct cache *cache);
 
 #endif /* _LIBPPCEMU_INTERNAL_CACHE_H */
